Change readByteArray signature to avoid too large arrays allocation

+++ benchmarks/src/commonMain/kotlin/BufferOps.kt
@@ -387,6 +387,6 @@ open class BufferReadNewByteArray: BufferRWBenchmarkBase() {
     @Benchmark
     fun benchmark(): ByteArray {
         buffer.write(inputArray)
-        return buffer.readByteArray(size.toLong())
+        return buffer.readByteArray(size)
     }
 }
